// bank-account-app-front/src/app/shared/models/events/customer-action.events.ts
export enum CustomerEvent {
    REFRESH_EVENT = "[Customer-API] REFRESH LIST DATA EVENT",
    CREATE_NEW_CUSTOMER_EVENT = "[Customer-API] CREATE NEW CUSTOMER EVENT",
    GET_ALL_CUSTOMERS_EVENT = "[Customer-API] GET ALL CUSTOMERS EVENT",
    SWITCH_CUSTOMER_STATE_EVENT = "[Customer-API] SWICTH STATE CUSTOMER  EVENT",
    UPDATE_CUSTOMER_EVENT = "[Customer-API] UPDATE CUSTOMER EVENT",
    ARCHIVE_CUSTOMER_EVENT = "[Customer-API] ARCHIVE CUSTOMER EVENT"
}