// bank-account-app-front/src/app/shared/models/events/account-action.events.ts
export enum AccountEvent {
    CREATE_ACCOUNT_EVENT = "[Account-API] CREATE ACCOUNT EVENT",
    REFRESH_EVENT = "[Account-API] REFRESH LIST OF ACCOUNTS",
    SWICTH_ACCOUNT_STATE_EVENT = "[Account-API] SWICTH ACCOUNT STATE EVENT",
    UPDATE_ACCOUNT_OVERDRAFT_OR_IRATE_EVENT = "[Account-API] UPDATE ACCOUNT OVERDRAFT OR IRATE EVENT",
}